--- app/src/main/java/com/example/yard/EventFragment.java
@@ -1,19 +1,0 @@
-package com.example.yard;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-public class EventFragment extends Fragment {
-
-    public EventFragment(){
-
-    }
-    @Nullable
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_events, container, false);
-    }
-}
