+++ app/src/main/java/com/example/yard/Fragments/BioTagFragment.java
@@ -1,0 +1,28 @@
+package com.example.yard.Fragments;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.yard.R;
+
+public class BioTagFragment extends Fragment {
+  public BioTagFragment() {
+    // Required empty public constructor
+  }
+
+  @Override
+  public void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+  }
+
+  @Override
+  public View onCreateView(
+      LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+    // Inflate the layout for this fragment
+    return inflater.inflate(R.layout.fragment_bio_tag, container, false);
+  }
+}
