--- app/src/main/java/com/example/yard/Post.java
@@ -1,219 +1,0 @@
-package com.example.yard;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.core.content.FileProvider;
-
-import android.Manifest;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.location.Address;
-import android.location.Geocoder;
-import android.location.Location;
-import android.net.Uri;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.google.android.gms.location.FusedLocationProviderClient;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.parse.ParseException;
-import com.parse.ParseFile;
-import com.parse.ParseUser;
-import com.parse.SaveCallback;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.List;
-import java.util.Locale;
-
-@RequiresApi(api = Build.VERSION_CODES.O)
-public class Post extends AppCompatActivity {
-    public static final int CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE = 45;
-    public static final int REQUEST_CODE = 100;
-    public static final String TAG = "Post";
-    private Button takePicture;
-    private Button submit;
-    private TextView etdescription;
-    private Button upload;
-    private File photoFile;
-    private ImageView image;
-    Button getLocation;
-    TextView userAddress;
-    Button removeLocation;
-    FusedLocationProviderClient fusedLocationProviderClient;
-    private String photoFileName = "photo.jpg";
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_post);
-        takePicture = findViewById(R.id.btnTakePicture);
-        submit = findViewById(R.id.btnSubmit);
-        upload = findViewById(R.id.btnUploadPicture);
-        etdescription = findViewById(R.id.tvDescription);
-        image = findViewById(R.id.ivImage);
-        userAddress = findViewById(R.id.tvLocationAddress);
-        userAddress.setText("");
-        getLocation = findViewById(R.id.btnGetLocation);
-        removeLocation = findViewById(R.id.btnRemoveLocation);
-        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
-
-        getLocation.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                getLastLocation();
-            }
-        });
-        takePicture.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                launchcamera();
-            }
-        });
-        submit.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String description = etdescription.getText().toString();
-                String userLocation = userAddress.getText().toString();
-
-
-                if (description.isEmpty()){
-                    Toast.makeText(Post.this, "Description cannot be empty", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                ParseUser currentUser = ParseUser.getCurrentUser();
-                savePost(description, userLocation, currentUser, photoFile);
-            }
-        });
-        removeLocation.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String userLocation = userAddress.getText().toString();
-                userAddress.setText("");
-
-            }
-        });
-    }
-
-    private void getLastLocation() {
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-            fusedLocationProviderClient.getLastLocation().addOnSuccessListener(new OnSuccessListener<Location>() {
-                @Override
-                public void onSuccess(Location location) {
-                    if (location != null){
-                        Geocoder geocoder = new Geocoder(Post.this, Locale.getDefault());
-                        List<Address> addresses = null;
-                        try {
-                            addresses = geocoder.getFromLocation(location.getLatitude(), location.getLongitude(), 1);
-                            userAddress.setText("" + addresses.get(0).getLocality() + ", " +addresses.get(0).getCountryName());
-                        } catch (IOException e) {
-                            e.printStackTrace();
-                        }
-
-                    } else {
-                        askPermission();
-                    }
-                }
-            });
-        }
-    }
-
-    private void askPermission() {
-        ActivityCompat.requestPermissions(Post.this, new String[]
-                {Manifest.permission.ACCESS_FINE_LOCATION}, REQUEST_CODE);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        if (requestCode == REQUEST_CODE){
-            if(grantResults.length> 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
-                getLastLocation();
-            }
-            else{
-                Toast.makeText(this, "Required Permission", Toast.LENGTH_SHORT).show();
-            }
-        }
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-    }
-
-    private void launchcamera() {
-        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        // Create a File reference for future access
-        photoFile = getPhotoFileUri(photoFileName);
-        // wrap File object into a content provider
-        // required for API >= 24
-        // See https://guides.codepath.com/android/S haring-Content-with-Intents#sharing-files-with-api-24-or-higher
-        Uri fileProvider = FileProvider.getUriForFile(Post.this, "com.codepath.fileprovider", photoFile);
-        intent.putExtra(MediaStore.EXTRA_OUTPUT, fileProvider);
-        // If you call star tActivityForResult() using an intent that no app can handle, your app will crash.
-        // So as long as the result is not null, it's safe to use the intent.
-        if (intent.resolveActivity(Post.this.getPackageManager()) != null) {
-            // Start the image capture intent to take photo
-            startActivityForResult(intent, CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE);
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE) {
-            if (resultCode == RESULT_OK) {
-                // by this point we have the camera photo on disk
-                Bitmap takenImage = BitmapFactory.decodeFile(photoFile.getAbsolutePath());
-                // RESIZE BITMAP, see section below
-                // Load the taken image into a preview
-                image.setImageBitmap(takenImage);
-            } else { // Result was a failure
-                Toast.makeText(this, "Picture wasn't taken!", Toast.LENGTH_SHORT).show();
-            }
-        }
-    }
-
-    public File getPhotoFileUri(String fileName) {
-        // Get safe storage directory for photos
-        // Use `getExternalFilesDir` on Context to access package-specific directories.
-        // This way, we don't need to request external read/write runtime permissions.
-        File mediaStorageDir = new File(getExternalFilesDir(Environment.DIRECTORY_PICTURES), TAG);
-        // Create the storage directory if it does not exist
-        if (!mediaStorageDir.exists() && !mediaStorageDir.mkdirs()){
-            Log.d(TAG, "failed to create directory");
-        }
-        // Return the file target for the photo based on filename
-        return new File(mediaStorageDir.getPath() + File.separator + fileName);
-    }
-
-    private void  savePost(String description, String userLocation, ParseUser currentUser, File photoFile) {
-        PostCreation postCreation = new PostCreation();
-        postCreation.setKeyDescription(description);
-        postCreation.setKeyLocation(userLocation);
-        if (image.getDrawable() != null) {
-            postCreation.setImage(new ParseFile(photoFile));
-        }
-        postCreation.setKeyUser(currentUser);
-        postCreation.saveInBackground(new SaveCallback() {
-            @Override
-            public void done(ParseException e) {
-                if (e != null){
-                    Log.e("error", "error while saving:" + e);
-                    Toast.makeText(Post.this, "error while saving", Toast.LENGTH_SHORT).show();
-                }
-                etdescription.setText("");
-                image.setImageResource(0);
-            }
-        });
-    }
-}
