+++ app/src/main/res/anim/from_bottom.xml
@@ -3,16 +3,17 @@
     android:fillAfter="true">
 
     <translate
+        android:duration="300"
         android:fromXDelta="100%"
-        android:toXDelta="0%"
-        android:duration="300" />
+        android:toXDelta="0%" />
     <scale
-        android:pivotY="50%"
         android:pivotX="50%"
-        android:toYScale="0.8"
-        android:toXScale="0.8" />
-    <alpha android:fromAlpha="0"
-        android:toAlpha="1"
-        android:duration="800" />
+        android:pivotY="50%"
+        android:toXScale="0.8"
+        android:toYScale="0.8" />
+    <alpha
+        android:duration="800"
+        android:fromAlpha="0"
+        android:toAlpha="1" />
 
 </set>
