+++ app/src/main/res/anim/rotate_close.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:fillAfter="true" >
+    android:fillAfter="true">
 
     <rotate
+        android:duration="300"
+        android:fromDegrees="45"
         android:pivotX="50%"
         android:pivotY="50%"
-        android:fromDegrees="45"
-        android:toDegrees="0"
-        android:duration="300" />
+        android:toDegrees="0" />
 
 </set>
