+++ app/src/main/res/anim/rotate_open.xml
@@ -2,10 +2,10 @@
 <set xmlns:android="http://schemas.android.com/apk/res/android"
     android:fillAfter="true">
     <rotate
+        android:duration="300"
+        android:fromDegrees="0"
         android:pivotX="50%"
         android:pivotY="50%"
-        android:fromDegrees="0"
-        android:toDegrees="45"
-        android:duration="300" />
+        android:toDegrees="45" />
 
 </set>
