+++ app/src/main/res/anim/to_bottom.xml
@@ -3,18 +3,19 @@
     android:fillAfter="true">
 
     <translate
+        android:duration="300"
         android:fromXDelta="0%"
-        android:toXDelta="100%"
-        android:duration="300" />
+        android:toXDelta="100%" />
     <scale
-        android:pivotY="50%"
-        android:pivotX="50%"
         android:fromXScale="0.8"
         android:fromYScale="0.8"
-        android:toYScale="0.8"
-        android:toXScale="0.8" />
-    <alpha android:fromAlpha="1"
-        android:toAlpha="0"
-        android:duration="150" />
+        android:pivotX="50%"
+        android:pivotY="50%"
+        android:toXScale="0.8"
+        android:toYScale="0.8" />
+    <alpha
+        android:duration="150"
+        android:fromAlpha="1"
+        android:toAlpha="0" />
 
 </set>
