+++ app/src/main/res/drawable-v24/button_background.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <stroke android:width="1dp"/>
-    <solid android:color="@android:color/transparent"/>
+    <stroke android:width="1dp" />
+    <solid android:color="@android:color/transparent" />
 </shape>
