+++ app/src/main/res/drawable/ic_log_out_sign_svgrepo_com.xml
@@ -3,22 +3,22 @@
     android:height="32dp"
     android:viewportWidth="32"
     android:viewportHeight="32">
-  <path
-      android:pathData="M21,18l0,-7l-10,-10l-10,10l0,20l20,0l0,-7"
-      android:strokeLineJoin="round"
-      android:strokeWidth="2"
-      android:fillColor="@color/white"
-      android:strokeColor="#000000"/>
-  <path
-      android:pathData="M10,21L31,21"
-      android:strokeLineJoin="round"
-      android:strokeWidth="2"
-      android:fillColor="#00000000"
-      android:strokeColor="@color/black"/>
-  <path
-      android:pathData="M25,15l6,6l-6,6"
-      android:strokeLineJoin="round"
-      android:strokeWidth="2"
-      android:fillColor="@color/white"
-      android:strokeColor="@color/black"/>
+    <path
+        android:fillColor="@color/white"
+        android:pathData="M21,18l0,-7l-10,-10l-10,10l0,20l20,0l0,-7"
+        android:strokeWidth="2"
+        android:strokeColor="#000000"
+        android:strokeLineJoin="round" />
+    <path
+        android:fillColor="#00000000"
+        android:pathData="M10,21L31,21"
+        android:strokeWidth="2"
+        android:strokeColor="@color/black"
+        android:strokeLineJoin="round" />
+    <path
+        android:fillColor="@color/white"
+        android:pathData="M25,15l6,6l-6,6"
+        android:strokeWidth="2"
+        android:strokeColor="@color/black"
+        android:strokeLineJoin="round" />
 </vector>
