+++ app/src/main/res/layout/fragment_bio.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BioFragment">
+    tools:context=".Fragments.BioFragment">
 
 
     <com.google.android.material.appbar.AppBarLayout
