+++ app/src/main/res/layout/fragment_bio_tag.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BioTagFragment">
+    tools:context=".Fragments.BioTagFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
