+++ app/src/main/res/menu/bottom_navigation.xml
@@ -14,5 +14,4 @@
         android:title="Bio" />
 
 
-
 </menu>
