+++ app/src/main/res/xml/fileprovider.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+<PreferenceScreen>
     <paths>
         <!-- See table below.  The external-files-path DOES NOT require external storage permissions. -->
         <external-files-path
